Make storybook tell 2 stories
