+++ .storybook/manager-head.html
@@ -1,4 +1,4 @@
-
+<!-- 
 <script src="https://www.gstatic.com/firebasejs/7.2.0/firebase-app.js"></script>
 <script src="https://www.gstatic.com/firebasejs/7.2.0/firebase-auth.js"></script>
 
@@ -19,4 +19,4 @@
       window.location.assign(`auth.html`)
     }
   });
-</script>
+</script> -->
