+++ src/components/PokerClient/Player/README.md
