+++ src/components/PokerClient/PlayingCard/cardElements/Rank2.jsx
