+++ src/components/PokerClient/PlayingCard/cardElements/Rank3.jsx
