+++ src/components/PokerClient/PlayingCard/cardElements/Rank6.jsx
