+++ src/components/PokerClient/PlayingCard/cardElements/Rank7.jsx
