+++ src/components/PokerClient/PlayingCard/cardElements/Rank8.jsx
