+++ src/components/PokerClient/PlayingCard/cardElements/RankT.jsx
