+++ src/stories/viewports.js
@@ -1,25 +1,32 @@
 const viewports = {
-    iPhone: {
-      name: 'iPhone 6/7/8',
-      styles: {
-        width: '375px',
-        height: '667px',
-      },
-    },
-    galaxyS5: {
-      name: 'Galaxy S5',
-      styles: {
-        width: '640px',
-        height: '360px',
-      },
-    },
-    ipadPortrait: {
-      name: 'iPad Portrait',
-      styles: {
-        width: '1024px',
-        height: '768px',
-      },
-    },
-  };
+  iPhone: {
+    name: "iPhone 6/7/8",
+    styles: {
+      width: "375px",
+      height: "667px"
+    }
+  },
+  tiny: {
+    name: "tiny",
+    styles: {
+      height: "600px",
+      width: "360px"
+    }
+  },
+  galaxyS5: {
+    name: "Galaxy S5",
+    styles: {
+      height: "640px",
+      width: "360px"
+    }
+  },
+  ipadPortrait: {
+    name: "iPad Portrait",
+    styles: {
+      width: "1024px",
+      height: "768px"
+    }
+  }
+};
 
 export default viewports;
